% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openserver-package.R
\name{createCOMReference}
\alias{createCOMReference}
\title{COM Reference call}
\usage{
createCOMReference(ref, className)
}
\arguments{
\item{ref}{the reference to pass}

\item{className}{the name of the class}
}
\description{
Call COM client reference function
}
\keyword{internal}
