% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openserver-class.R
\name{DoCmd}
\alias{DoCmd}
\title{Send an OpenServer command}
\usage{
DoCmd(object, command)
}
\arguments{
\item{command}{command to send}

\item{object}{OpenServer object}
}
\description{
Send an OpenServer command
}
\section{R6 Usage}{
 $DoCmd()
}

