% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{DoGAPFunc}
\alias{DoGAPFunc}
\title{Run a GAP function}
\usage{
DoGAPFunc(object, Gv)
}
\arguments{
\item{Gv}{GAP variable}

\item{object}{OpenServer object}
}
\description{
Run a GAP function
}
\section{R6 Usage}{
 $DoGAPFunc()
}

