% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openserver-class.R
\name{DoGet}
\alias{DoGet}
\title{Get a value from an OpenServer variable}
\usage{
DoGet(object, variable)
}
\arguments{
\item{variable}{OpenServer variable}

\item{object}{OpenServer object}
}
\description{
Get a value from an OpenServer variable
}
\section{R6 Usage}{
 $DoGet()
}

