% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openserver-class.R
\name{DoSet}
\alias{DoSet}
\title{Set a value to an OpenServer variable}
\usage{
DoSet(object, variable, set_value)
}
\arguments{
\item{variable}{OpenServer variable}

\item{set_value}{value to set to OpenServer variable}

\item{object}{OpenServer object}
}
\description{
Set a value to an OpenServer variable
}
\section{R6 Usage}{
 $DoSet()
}

