% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{DoSlowCmd}
\alias{DoSlowCmd}
\title{Send a command with a delay}
\usage{
DoSlowCmd(object, command)
}
\arguments{
\item{command}{OpenServer command}

\item{object}{OpenServer object}
}
\description{
Send a command with a delay
}
\section{R6 Usage}{
 $DoSlowCmd()
}

