% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{OpenServer}
\alias{OpenServer}
\title{OpenServer object}
\value{
Returns R6 object of class OpenServer
}
\description{
A class to create OpenServer objects
}
\section{Constructor}{
 OpenServer$new(server = NULL)
}

\section{Public Methods}{

\tabular{ll}{
\strong{Parameters Methods} \tab \strong{Link} \cr
\code{DoCmd()} \tab \code{\link{DoCmd}} \cr
\code{DoGet()} \tab \code{\link{DoGet}} \cr
\code{DoSet()} \tab \code{\link{DoSet}} \cr
\code{DoSlowCmd()} \tab \code{\link{DoSlowCmd}} \cr
\code{DoGAPFunc()} \tab \code{\link{DoGAPFunc}} \cr

}
}

