% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{openserver_to_character}
\alias{openserver_to_character}
\title{Convert from OpenServer text object to character}
\usage{
openserver_to_character(x)
}
\arguments{
\item{x}{an Openserver object}
}
\description{
Convert an OpenServer vector (string separated by '|') to a vector of character
}
