% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataobjects.R
\name{openserver_to_date}
\alias{openserver_to_date}
\title{Convert from OpenServer text object to date}
\usage{
openserver_to_date(x, format = "\%m/\%d/\%Y")
}
\arguments{
\item{x}{an Openserver object}

\item{format}{use the as.Date formats}
}
\description{
Convert an OpenServer vector (string separated by '|') to a vector of dates
}
