% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{openserver_to_double}
\alias{openserver_to_double}
\title{Convert from OpenServer text object to double}
\usage{
openserver_to_double(x)
}
\arguments{
\item{x}{an Openserver object}
}
\description{
Convert an OpenServer vector (string separated by '|') to a vector of double
}
