% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{openserver_to_integer}
\alias{openserver_to_integer}
\title{Convert from OpenServer text object to integer}
\usage{
openserver_to_integer(x)
}
\arguments{
\item{x}{an Openserver object}
}
\description{
Convert an OpenServer vector (string separated by '|') to a vector of integers
}
