% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataobjects.R
\name{rDoGet}
\alias{rDoGet}
\title{Function that converts OpenServer raw vectors to R vectors}
\usage{
rDoGet(oserver, cmd)
}
\arguments{
\item{oserver}{openserver object}

\item{cmd}{openserver command or variable}
}
\description{
Get openserver command and convert it to R object
}
