% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openserver-package.R
\docType{package}
\name{rOpenserver-package}
\alias{rOpenserver}
\alias{rOpenserver-package}
\title{rOpenserver: COM interface to OpenServer written in R}
\description{
Provide a COM interface to R so it can talk to Petex's IPM 
    applications Prosper, GAP, MBAL via OpenServer.
}
\keyword{internal}
