% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openserver-functions.R
\name{setOpenServer}
\alias{setOpenServer}
\title{OpenServer constructor}
\usage{
setOpenServer(server = NULL)
}
\arguments{
\item{server}{pass an existing OpenServer instance}
}
\description{
Openserver constructor that saves typing
\code{.OpenServer$new()}.
}
\seealso{
Other server-side items: 
\code{\link{.OpenServer}},
\code{\link{OpenServerR6}},
\code{\link{newOpenServer}()},
\code{\link{openserver}()}
}
\concept{server-side items}
