% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openserver-functions.R
\name{openserver}
\alias{openserver}
\title{Openserver constructor without using `$new()`}
\usage{
openserver(server = NULL)
}
\arguments{
\item{server}{an existing OpenServer instance}
}
\description{
Another Openserver constructor. Same as
}
\seealso{
Other server-side items: 
\code{\link{.OpenServer}},
\code{\link{newOpenServer}()},
\code{\link{setOpenServer}()}
}
\concept{server-side items}
