% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{OpenServer}
\alias{OpenServer}
\alias{createCOMReference}
\title{Openserver function
WIll call class}
\usage{
OpenServer(server = NULL)
}
\arguments{
\item{server}{an existing OpenServer instance}
}
\description{
Openserver function
WIll call class
}
