% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{openserver}
\alias{openserver}
\title{Another Openserver constructor without using `$new()`}
\usage{
openserver(server = NULL)
}
\arguments{
\item{server}{an existing OpenServer instance}
}
\description{
Call class
}
