% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openserver-functions.R
\name{openserver}
\alias{openserver}
\title{OpenServer constructor}
\usage{
openserver(server = NULL)
}
\arguments{
\item{server}{an existing OpenServer instance}
}
\description{
Openserver constructor that saves typing
\code{.OpenServer$new()}. This is used in rProsper.
}
\seealso{
Other server-side items: 
\code{\link{.OpenServer}},
\code{\link{OpenServerR6}},
\code{\link{newOpenServer}()},
\code{\link{setOpenServer}()}
}
\concept{server-side items}
