% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{read_openserver_status}
\alias{read_openserver_status}
\title{Read the status of OpenServer and IPM objects}
\usage{
read_openserver_status(verbose = FALSE)
}
\arguments{
\item{verbose}{indicate if we want live printing}
}
\description{
Create a dataframe with the count of OpenServer and IPM objects
}
