% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{setOpenServer}
\alias{setOpenServer}
\title{Openserver constructor
Call class}
\usage{
setOpenServer(server = NULL)
}
\arguments{
\item{server}{an existing OpenServer instance}
}
\description{
Openserver constructor
Call class
}
