% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openserver-functions.R
\name{setOpenServer}
\alias{setOpenServer}
\title{Another Openserver constructor without using `$new()`}
\usage{
setOpenServer(server = NULL)
}
\arguments{
\item{server}{an existing OpenServer instance}
}
\description{
Call class
}
\seealso{
Other server-side items: 
\code{\link{.OpenServer}},
\code{\link{newOpenServer}()}
}
\concept{server-side items}
